requirements.txt
mecab-python3==1.0.1
gensim==3.8.3
scikit-learn==0.22.1
mojimoji==0.0.9